requirements.txt
pyyaml
pandas
xlsxwriter
progressbar2
requests
beautifulsoup4